requirements.txt
astroid==1.5.3
autopep8==1.3.3
certifi==2017.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
Django==2.0
django-debug-toolbar==1.9.1
djangorestframework==3.7.3
idna==2.6
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
mod-wsgi==4.5.22+ap24vc14
mysqlclient==1.3.12
packaging==16.8
pip-review==1.0
pycodestyle==2.3.1
pylint==1.7.5
pylint-django==0.7.2
pylint-plugin-utils==0.2.6
pyparsing==2.2.0
python-dotenv==0.7.1
pytz==2017.3
requests==2.18.4
six==1.11.0
sqlparse==0.2.4
urllib3==1.22
wrapt==1.10.11
